Refactor History/Hashing unwind stages.
