+++ crates/stages/src/stages/index_storage_history.rs
@@ -1,15 +1,8 @@
 use crate::{ExecInput, ExecOutput, Stage, StageError, StageId, UnwindInput, UnwindOutput};
-use reth_db::{
-    cursor::{DbCursorRO, DbCursorRW},
-    database::{Database, DatabaseGAT},
-    models::storage_sharded_key::StorageShardedKey,
-    tables,
-    transaction::{DbTx, DbTxMut, DbTxMutGAT},
-    TransitionList,
-};
-use reth_primitives::{Address, TransitionId, H256};
+use reth_db::{database::Database, models::TransitionIdAddress};
+use reth_primitives::Address;
 use reth_provider::Transaction;
-use std::{collections::BTreeMap, fmt::Debug};
+use std::fmt::Debug;
 use tracing::*;
 
 /// The [`StageId`] of the storage history indexing stage.
@@ -74,94 +67,32 @@ impl<DB: Database> Stage<DB> for IndexStorageHistoryStage {
         let from_transition_rev = tx.get_block_transition(input.unwind_to)?;
         let to_transition_rev = tx.get_block_transition(input.stage_progress)?;
 
-        let mut cursor = tx.cursor_write::<tables::StorageHistory>()?;
+        tx.unwind_storage_history_indices(
+            TransitionIdAddress((from_transition_rev, Address::zero()))..
+                TransitionIdAddress((to_transition_rev, Address::zero())),
+        )?;
 
-        let storage_changesets = tx
-            .cursor_read::<tables::StorageChangeSet>()?
-            .walk(Some((from_transition_rev, Address::zero()).into()))?
-            .take_while(|res| {
-                res.as_ref().map(|(k, _)| k.transition_id() < to_transition_rev).unwrap_or_default()
-            })
-            .collect::<Result<Vec<_>, _>>()?;
-        let last_indices = storage_changesets
-            .into_iter()
-            // reverse so we can get lowest transition id where we need to unwind account.
-            .rev()
-            // fold all storages and get last transition index
-            .fold(
-                BTreeMap::new(),
-                |mut accounts: BTreeMap<(Address, H256), u64>, (index, storage)| {
-                    // we just need address and lowest transition id.
-                    accounts.insert((index.address(), storage.key), index.transition_id());
-                    accounts
-                },
-            );
-        for ((address, storage_key), rem_index) in last_indices {
-            let shard_part =
-                unwind_storage_history_shards::<DB>(&mut cursor, address, storage_key, rem_index)?;
-
-            // check last shard_part, if present, items needs to be reinserted.
-            if !shard_part.is_empty() {
-                // there are items in list
-                tx.put::<tables::StorageHistory>(
-                    StorageShardedKey::new(address, storage_key, u64::MAX),
-                    TransitionList::new(shard_part)
-                        .expect("There is at least one element in list and it is sorted."),
-                )?;
-            }
-        }
         Ok(UnwindOutput { stage_progress: input.unwind_to })
     }
 }
 
-/// Unwind all history shards. For boundary shard, remove it from database and
-/// return last part of shard with still valid items. If all full shard were removed, return list
-/// would be empty but this does not mean that there is none shard left but that there is no
-/// splitted shards.
-pub fn unwind_storage_history_shards<DB: Database>(
-    cursor: &mut <<DB as DatabaseGAT<'_>>::TXMut as DbTxMutGAT<'_>>::CursorMut<
-        tables::StorageHistory,
-    >,
-    address: Address,
-    storage_key: H256,
-    transition_id: TransitionId,
-) -> Result<Vec<usize>, StageError> {
-    let mut item = cursor.seek_exact(StorageShardedKey::new(address, storage_key, u64::MAX))?;
-
-    while let Some((storage_sharded_key, list)) = item {
-        // there is no more shard for address
-        if storage_sharded_key.address != address ||
-            storage_sharded_key.sharded_key.key != storage_key
-        {
-            // there is no more shard for address and storage_key.
-            break
-        }
-        cursor.delete_current()?;
-        // check first item and if it is more and eq than `transition_id` delete current
-        // item.
-        let first = list.iter(0).next().expect("List can't empty");
-        if first >= transition_id as usize {
-            item = cursor.prev()?;
-            continue
-        } else if transition_id <= storage_sharded_key.sharded_key.highest_transition_id {
-            // if first element is in scope whole list would be removed.
-            // so at least this first element is present.
-            return Ok(list.iter(0).take_while(|i| *i < transition_id as usize).collect::<Vec<_>>())
-        } else {
-            return Ok(list.iter(0).collect::<Vec<_>>())
-        }
-    }
-    Ok(Vec::new())
-}
 #[cfg(test)]
 mod tests {
 
+    use std::collections::BTreeMap;
+
     use super::*;
     use crate::test_utils::{TestTransaction, PREV_STAGE_ID};
-    use reth_db::models::{
-        storage_sharded_key::NUM_OF_INDICES_IN_SHARD, ShardedKey, TransitionIdAddress,
+    use reth_db::{
+        models::{
+            storage_sharded_key::{StorageShardedKey, NUM_OF_INDICES_IN_SHARD},
+            ShardedKey, TransitionIdAddress,
+        },
+        tables,
+        transaction::DbTxMut,
+        TransitionList,
     };
-    use reth_primitives::{hex_literal::hex, StorageEntry, H160, U256};
+    use reth_primitives::{hex_literal::hex, StorageEntry, H160, H256, U256};
 
     const ADDRESS: H160 = H160(hex!("0000000000000000000000000000000000000001"));
     const STORAGE_KEY: H256 =
